requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.8
cryptography==43.0.1
decorator==5.2.1
distlib==0.3.8
distro==1.9.0
dnspython==2.7.0
docutils==0.21.2
executing==2.2.0
fastapi==0.115.12
filelock==3.16.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.1.0
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jedi==0.19.2
jeepney==0.8.0
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
keyring==25.4.1
langchain==0.3.21
langchain-core==0.3.49
langchain-text-splitters==0.3.7
langsmith==0.3.19
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
nh3==0.2.18
numpy==2.2.4
openai==1.68.2
orjson==3.10.16
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pipenv==2024.0.3
pkginfo==1.10.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
pymongo==4.11.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.8.1
SecretStorage==3.3.3
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
stack-data==0.6.3
starlette==0.46.1
tenacity==9.0.0
tqdm==4.67.1
traitlets==5.14.3
twine==5.1.1
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.26.5
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
wheel==0.44.0
zipp==3.20.2
zstandard==0.23.0
